fix(technical-config): address comparison review findings
